# Improve performance (IPC)