--- a/.gitignore
+++ b/.gitignore
@@ -15,3 +15,37 @@ Cargo.lock
 # Ignore Gradle build output directory
 build
 .idea
+
+
+## Java
+
+# Compiled class file
+*.class
+
+# Log file
+*.log
+
+# BlueJ files
+*.ctxt
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
+
+# Package Files #
+*.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+
+# IDE generated
+.vscode
+.classpath
+.settings
+.project
+bin