--- a/flight-client/src/main/java/org/m4d/adp/flight_client/MyFlightClient.java
+++ b/flight-client/src/main/java/org/m4d/adp/flight_client/MyFlightClient.java
@@ -58,6 +58,7 @@ public class MyFlightClient {
         try {
             s.close();
             client.close();
+            allocator.close();
         } catch (Exception e) {
             e.printStackTrace();
         }