--- a/flight-server/src/main/java/org/m4d/adp/flight_server/NoCopyByteArrayOutputStream.java
+++ b/flight-server/src/main/java/org/m4d/adp/flight_server/NoCopyByteArrayOutputStream.java
@@ -1,0 +1,14 @@
+package org.m4d.adp.flight_server;
+import java.io.ByteArrayOutputStream;
+
+class NoCopyByteArrayOutputStream extends ByteArrayOutputStream {
+    public NoCopyByteArrayOutputStream() {
+         super(); 
+    }
+    public NoCopyByteArrayOutputStream(int size) {
+        super(size);
+    }
+    public byte[] getBuffer() {
+        return buf; 
+    }
+}