--- a/flight-server/src/main/java/org/m4d/adp/flight_server/RelayProducer.java
+++ b/flight-server/src/main/java/org/m4d/adp/flight_server/RelayProducer.java
@@ -19,52 +19,52 @@ import org.m4d.adp.allocator.WasmAllocationFactory;
 import org.m4d.adp.transform.TransformInterface;
 
 import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.nio.ByteBuffer;
 import java.nio.channels.Channels;
 
-/** Flight Producer for flight server that serves as relay to another
- *  flight server (e.g. the flight server with ExampleProducer).
- *  Depending on the transform boolean, this server may transform
- *  the first column with a constant column.
+/**
+ * Flight Producer for flight server that serves as relay to another flight
+ * server (e.g. the flight server with ExampleProducer). Depending on the
+ * transform boolean, this server may transform the first column with a constant
+ * column.
  */
-public class RelayProducer extends NoOpFlightProducer {
+public class RelayProducer extends NoOpFlightProducer implements AutoCloseable {
     BufferAllocator allocator;
-    private boolean transform = false;
+    private final boolean zeroCopy;
+    private final boolean transform;
     private final FlightClient client;
     private final Location location;
 
-    public RelayProducer(Location location, Location remote_location,
-                          BufferAllocator allocator, boolean transform) {
+    public RelayProducer(Location location, Location remote_location, BufferAllocator allocator, boolean transform,
+            boolean zeroCopy) {
         this.transform = transform;
+        this.zeroCopy = zeroCopy;
         this.allocator = allocator;
         this.location = location;
-        this.client = FlightClient.builder()
-                .allocator(allocator)
-                .location(remote_location)
-                .build();
+        this.client = FlightClient.builder().allocator(allocator).location(remote_location).build();
     }
 
-    private byte[] WASMTransformByteArray(long instance_ptr, byte[] recordBatchByteArray) {
-        long length = recordBatchByteArray.length;
+    private byte[] WASMTransformByteArray(long instance_ptr, byte[] recordBatchByteArray, int size) {
         // Allocate a block in wasm memory and copy the byte array to this block
-        long allocatedAddress = AllocatorInterface.wasmAlloc(instance_ptr, length);
+        long allocatedAddress = AllocatorInterface.wasmAlloc(instance_ptr, size);
         long wasm_mem_address = AllocatorInterface.wasmMemPtr(instance_ptr);
-        ByteBuffer buffer = MemoryUtil.directBuffer(allocatedAddress + wasm_mem_address, (int) length);
-        buffer.put(recordBatchByteArray);
-        
-        // Transform the vector schema root that is represented as a byte array in `allocatedAddress` memory address with length `length`
-        // The function returns a tuple of `(address, lenght)` of as byte array that represents the transformed vector schema root 
-        long transformed_bytes_tuple = TransformInterface.TransformationIPC(instance_ptr, allocatedAddress, length);
-        
+        ByteBuffer buffer = MemoryUtil.directBuffer(allocatedAddress + wasm_mem_address, size);
+        buffer.put(recordBatchByteArray, 0, size);
+
+        // Transform the vector schema root that is represented as a byte array in
+        // `allocatedAddress` memory address with length `length`
+        // The function returns a tuple of `(address, lenght)` of as byte array that
+        // represents the transformed vector schema root
+        long transformed_bytes_tuple = TransformInterface.TransformationIPC(instance_ptr, allocatedAddress, size);
+
         // Get the byte array from the memory address
         long transformed_bytes_address = TransformInterface.GetFirstElemOfTuple(instance_ptr, transformed_bytes_tuple);
         long transformed_bytes_len = TransformInterface.GetSecondElemOfTuple(instance_ptr, transformed_bytes_tuple);
         wasm_mem_address = AllocatorInterface.wasmMemPtr(instance_ptr);
-        ByteBuffer transformed_buffer = MemoryUtil.directBuffer(transformed_bytes_address + wasm_mem_address, (int) transformed_bytes_len);
+        ByteBuffer transformed_buffer = MemoryUtil.directBuffer(transformed_bytes_address + wasm_mem_address,
+                (int) transformed_bytes_len);
         byte[] transformedRecordBatchByteArray = new byte[(int) transformed_bytes_len];
         transformed_buffer.get(transformedRecordBatchByteArray);
         // Deallocate transformed bytes
@@ -73,89 +73,106 @@ public class RelayProducer extends NoOpFlightProducer {
     }
 
     @Override
-    public void getStream(CallContext context, Ticket ticket,
-                          ServerStreamListener listener) {
-        FlightStream s = client.getStream(ticket);
-        VectorSchemaRoot root_in;
-        VectorSchemaRoot root_out = null;
-        VectorLoader loader = null;
-        VectorUnloader unloader = null;
-        ArrowStreamReader reader = null;
-
-        WasmAllocationFactory wasmAllocationFactory = new WasmAllocationFactory();
-        long instance_ptr = wasmAllocationFactory.wasmInstancePtr();
-        boolean first = true;
-        while (s.next()) {
-            root_in = s.getRoot();
-            if (first) {
-                root_out = VectorSchemaRoot.create(root_in.getSchema(), allocator);
-                loader = new VectorLoader(root_out);
-                listener.setUseZeroCopy(false);
-                listener.start(root_out);
-                first = false;
-            }
-            if (transform) {
-                ByteArrayOutputStream out = new ByteArrayOutputStream();
-                try (
-                        ArrowStreamWriter writer = new ArrowStreamWriter(root_in, null, Channels.newChannel(out));
-                ) {
-                    // Write the vector schema root to an IPC stream
-                    writer.start();
-                    writer.writeBatch();
-                    writer.end();
-                    // Get a byte array of the vector schema root
-                    byte[] recordBatchByteArray = out.toByteArray();
-                    byte[] transformedRecordBatchByteArray = WASMTransformByteArray(instance_ptr, recordBatchByteArray);
-                    // Read the byte array to get the transformed vector schema root
-                    reader = new ArrowStreamReader(new ByteArrayInputStream(transformedRecordBatchByteArray), allocator);
-                    VectorSchemaRoot readBatch = reader.getVectorSchemaRoot();
-                    reader.loadNextBatch();
-                    root_in = readBatch;
-
-                } catch (Exception e) {
-                    root_in = null;
-                }
-            }
-
-            unloader = new VectorUnloader(root_in);
-            loader.load(unloader.getRecordBatch());
-            try {
-                if (reader != null) {
-                    reader.close();
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            listener.putNext();
-        }
-        listener.completed();
-
+    public void getStream(CallContext context, Ticket ticket, ServerStreamListener listener) {
         try {
-            s.close();
-            wasmAllocationFactory.close();
+            if (transform) {
+                getStreamTransform(context, ticket, listener);
+            } else {
+                getStreamNoTransform(context, ticket, listener);
+            }
+            System.out.println("relay producer - completed");
         } catch (Exception e) {
+            listener.error(e);
             e.printStackTrace();
+            System.out.println("relay producer - failed");
         }
-        System.out.println("relay producer - completed");
+    }
+
+    public void getStreamTransform(CallContext context, Ticket ticket, ServerStreamListener listener) throws Exception {
+        try (FlightStream stream = client.getStream(ticket);
+                WasmAllocationFactory wasmAllocationFactory = new WasmAllocationFactory()) {
+            long instance_ptr = wasmAllocationFactory.wasmInstancePtr();
+            VectorLoader loader = null;
+            while (stream.next()) {
+                try (NoCopyByteArrayOutputStream out = new NoCopyByteArrayOutputStream()) {
+                    // Write the input vector schema root to the byte array output stream in IPC
+                    // format
+                    try (ArrowStreamWriter writer = new ArrowStreamWriter(stream.getRoot(), null,
+                            Channels.newChannel(out))) {
+                        writer.start();
+                        writer.writeBatch();
+                        writer.end();
+                    }
+
+                    // Transform the original record batch
+                    byte[] transformedRecordBatchByteArray = WASMTransformByteArray(instance_ptr, out.getBuffer(),
+                            out.size());
+
+                    // Read the byte array to get the transformed batch
+                    try (ArrowStreamReader reader = new ArrowStreamReader(
+                            new ByteArrayInputStream(transformedRecordBatchByteArray), allocator)) {
+                        VectorSchemaRoot transformedVSR = reader.getVectorSchemaRoot();
+                        reader.loadNextBatch();
+                        VectorUnloader unloader = new VectorUnloader(transformedVSR);
+
+                        // First time initialization
+                        if (loader == null) {
+                            VectorSchemaRoot root_out = VectorSchemaRoot.create(transformedVSR.getSchema(), allocator);
+                            loader = new VectorLoader(root_out);
+                            listener.setUseZeroCopy(zeroCopy);
+                            listener.start(root_out);
+                        }
+                        loader.load(unloader.getRecordBatch());
+                        listener.putNext();
+                    }
+                }
+            }
+            listener.completed();
+        }
+    }
+
+    public void getStreamNoTransform(CallContext context, Ticket ticket, ServerStreamListener listener)
+            throws Exception {
+        VectorSchemaRoot root_out = null;
+        VectorUnloader unloader = null;
+        VectorLoader loader = null;
+        FlightStream stream = client.getStream(ticket);
+        try {
+            while (stream.next()) {
+                // First time initialization
+                if (unloader == null) {
+                    VectorSchemaRoot root = stream.getRoot();
+                    unloader = new VectorUnloader(root);
+                    root_out = VectorSchemaRoot.create(root.getSchema(), allocator);
+                    loader = new VectorLoader(root_out);
+                    listener.setUseZeroCopy(zeroCopy);
+                    listener.start(root_out);
+                }
+
+                loader.load(unloader.getRecordBatch());
+                listener.putNext();
+            }
+            listener.completed();
+        } finally {
+            stream.close();
+            if (root_out != null) {
+                root_out.close();
+            }
+        }
     }
 
     @Override
-    public FlightInfo getFlightInfo(FlightProducer.CallContext context,
-                                    FlightDescriptor descriptor) {
+    public FlightInfo getFlightInfo(FlightProducer.CallContext context, FlightDescriptor descriptor) {
         Preconditions.checkArgument(descriptor.isCommand());
-        //Perf exec = Perf.parseFrom(descriptor.getCommand());
+        // Perf exec = Perf.parseFrom(descriptor.getCommand());
 
-        final Schema pojoSchema = new Schema(ImmutableList.of(
-                Field.nullable("a", Types.MinorType.BIGINT.getType()),
+        final Schema pojoSchema = new Schema(ImmutableList.of(Field.nullable("a", Types.MinorType.BIGINT.getType()),
                 Field.nullable("b", Types.MinorType.BIGINT.getType()),
                 Field.nullable("c", Types.MinorType.BIGINT.getType()),
-                Field.nullable("d", Types.MinorType.BIGINT.getType())
-        ));
+                Field.nullable("d", Types.MinorType.BIGINT.getType())));
 
         int streamCount = 1;
-        PerfOuterClass.Token token = PerfOuterClass.Token.newBuilder()
-                .build();
+        PerfOuterClass.Token token = PerfOuterClass.Token.newBuilder().build();
         final Ticket ticket = new Ticket(token.toByteArray());
 
         List<FlightEndpoint> endpoints = new ArrayList<>();
@@ -164,5 +181,10 @@ public class RelayProducer extends NoOpFlightProducer {
         }
 
         return new FlightInfo(pojoSchema, descriptor, endpoints, -1, -1);
+    }
+
+    @Override
+    public void close() throws Exception {
+        this.client.close();        
     }
 }