--- a/flight-server/src/main/java/org/m4d/adp/flight_server/RelayProducer.java
+++ b/flight-server/src/main/java/org/m4d/adp/flight_server/RelayProducer.java
@@ -185,6 +185,6 @@ public class RelayProducer extends NoOpFlightProducer implements AutoCloseable {
 
     @Override
     public void close() throws Exception {
-        this.client.close();        
+        this.client.close();
     }
 }