--- a/java-wasm-interface/src/main/java/org/m4d/adp/allocator/WasmAllocationFactory.java
+++ b/java-wasm-interface/src/main/java/org/m4d/adp/allocator/WasmAllocationFactory.java
@@ -11,7 +11,7 @@ import org.apache.arrow.memory.ReferenceManager;
  */
 public class WasmAllocationFactory implements AllocationManager.Factory, AutoCloseable{
     // Pointer to an instance of the WASM module that is used to allocate and deallocate memory.
-    long instancePtr = AllocatorInterface.wasmInstance("../wasm-modules/allocator/target/wasm32-unknown-unknown/release/alloc.wasm");
+    long instancePtr = AllocatorInterface.wasmInstance("../wasm-modules/allocator/target/wasm32-wasi/release/alloc.wasm");
 
     @Override
     public AllocationManager create(BufferAllocator accountingAllocator, long size) {