--- a/wasm-modules/allocator/README.md
+++ b/wasm-modules/allocator/README.md
@@ -2,4 +2,8 @@
 
 Run the following command to get a .wasm file:
 
-`cargo build --target wasm32-unknown-unknown --release`
+`cargo build --target wasm32-wasi --release`
+
+Optimize with:
+
+`wasm-opt -O2 target/wasm32-wasi/release/alloc.wasm -o target/wasm32-wasi/release/alloc.wasm`