--- a/wasm_interface/Cargo.toml
+++ b/wasm_interface/Cargo.toml
@@ -12,4 +12,5 @@ crate-type = ["cdylib"]
 libc = "0.2.7"
 jni = "0.19.0"
 wasmer = "2.0"
-arrow = "4.2.0"
+wasmer-wasi = "2.0"
+# arrow = "4.2.0"