Performance optimizations for the IPC approach. 
All results are just the first run on my own laptop (i.e., not proper benchmarking but gives some indication).

Without this PR:
```
Time spent traversing dataset: 204.623 seconds
Throughput: 78.19MB/sec
```

Changing to wasm-wasi:
```
Time spent traversing dataset: 100.394 seconds
Throughput: 159.37MB/sec
```

Other optimizations (remove the copy when taking the bytes from output stream, optimize the wasm code with wasm-opt, upgrade arrow and enable simd):
```
Time spent traversing dataset: 87.618 seconds
Throughput: 182.61MB/sec
``` 

For completion result without transformation:
```
Time spent traversing dataset: 37.132 seconds
Throughput: 430.90MB/sec
```