@roee88 2021-08-12
> What is the usage of `zeroCopy`?

I didn't use it, just made it configurable for the relay (was previously always calling `listener.setUseZeroCopy(false);`). 
I'm not sure when setting it to true is allowed.